Refactor MainMenu builder to use move REST requests to IntentServices.
This is an effort to move these items off the UI thread so that the
app plays by Android best practices.  MainMenu was the first
Each of the necessary REST calls will get a IntentService for the
time being.  The various Adapters should implement a Handler for
the message that is returned.